requirements.txt
# Core dependencies
google-generativeai>=0.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# File handling
python-docx>=1.0.0
PyPDF2>=3.0.0

# Embeddings and vector stores (for Career Pathway Recommender)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
# Alternative: pinecone-client>=2.2.0
# Alternative: chromadb>=0.4.0

# Evaluation and metrics
scikit-learn>=1.3.0

# API and web
requests>=2.31.0
aiohttp>=3.9.0

# Environment management
python-dotenv>=1.0.0

# Visualization (for dashboards)
plotly>=5.17.0
dash>=2.14.0
streamlit>=1.28.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# Optional: Azure OpenAI
# azure-identity>=1.14.0
# azure-ai-openai>=1.0.0

# Optional: Bias detection
# detoxify>=0.5.0
